# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the API for the Agent Memory service.",
        "title": "Agent Memory API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/facts/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Retrieve a specific fact by its ID, optionally including original sources.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "facts"
                ],
                "summary": "Get Fact",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Fact ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "boolean",
                        "description": "Include original sources",
                        "name": "include_sources",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/memory.Fact"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.apiError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/api.apiError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/api.apiError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.apiError"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Update the text and source of an existing fact.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "facts"
                ],
                "summary": "Update Fact",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Fact ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update Body",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/memory.FactUpdateBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/memory.Fact"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.apiError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/api.apiError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/api.apiError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.apiError"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Delete facts by ID or a list of IDs.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "facts"
                ],
                "summary": "Delete Facts",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Fact ID (optional if provided in body)",
                        "name": "id",
                        "in": "path"
                    },
                    {
                        "description": "Delete Body",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/memory.FactDeleteBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"status\": \"deleted\"}",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.apiError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/api.apiError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.apiError"
                        }
                    }
                }
            }
        },
...
        "/memory/contextual": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Process input through the contextual smart pipeline to retrieve relevant facts and messages.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "memory"
                ],
                "summary": "Process Contextual Memory",
                "parameters": [
                    {
                        "description": "Memory Input",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/memory.MemoryInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/memory.MemoryOutput"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.apiError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/api.apiError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.apiError"
                        }
                    }
                }
            }
        },
        "/memory/factual": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Add new facts to the memory engine.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "memory"
                ],
                "summary": "Add Factual Memory",
                "parameters": [
                    {
                        "description": "Factual Input",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/memory.FactualInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/memory.EvaluateResult"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.apiError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/api.apiError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.apiError"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "api.apiError": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
...
        "memory.ConversationMessage": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "event_id": {
                    "type": "string"
                },
                "kind": {
                    "$ref": "#/definitions/memory.SourceKind"
                },
                "session_id": {
                    "type": "string"
                },
                "source_id": {
                    "type": "string"
                }
            }
        },
        "memory.EvaluateResult": {
            "type": "object",
            "properties": {
                "facts_to_delete": {
                    "description": "IDs of facts to remove.",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "facts_to_return": {
                    "description": "Relevant existing + new facts to return to the agent.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/memory.Fact"
                    }
                },
                "facts_to_store": {
                    "description": "New facts to insert.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/memory.Fact"
                    }
                },
                "facts_to_update": {
                    "description": "Existing facts contradicted by higher/equal trust source.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/memory.Fact"
                    }
                }
            }
        },
        "memory.Fact": {
            "type": "object",
            "properties": {
                "account_id": {
                    "type": "string"
                },
                "agent_id": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "embedding": {
                    "type": "array",
                    "items": {
                        "type": "number"
                    }
                },
                "id": {
                    "type": "string"
                },
                "kind": {
                    "$ref": "#/definitions/memory.FactKind"
                },
                "session_id": {
                    "description": "Null means agent-level or account-level scope.",
                    "type": "string"
                },
                "source_id": {
                    "description": "The source that produced this fact.",
                    "type": "string"
                },
                "text": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "memory.FactDeleteBody": {
            "type": "object",
            "properties": {
                "fact_ids": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "source": {
                    "$ref": "#/definitions/memory.SourceKind"
                }
            }
        },
        "memory.FactKind": {
            "type": "string",
            "enum": [
                "KNOWLEDGE",
                "RULE",
                "PREFERENCE"
            ],
            "x-enum-comments": {
                "FACT_KIND_KNOWLEDGE": "Factual information. Retrieved only when semantically relevant.",
                "FACT_KIND_PREFERENCE": "Soft weight for decisions. Retrieved when relevant to choices.",
                "FACT_KIND_RULE": "Always included in retrieval. User-defined rules."
            },
            "x-enum-descriptions": [
                "Factual information. Retrieved only when semantically relevant.",
                "Always included in retrieval. User-defined rules.",
                "Soft weight for decisions. Retrieved when relevant to choices."
            ],
            "x-enum-varnames": [
                "FACT_KIND_KNOWLEDGE",
                "FACT_KIND_RULE",
                "FACT_KIND_PREFERENCE"
            ]
        },
        "memory.FactUpdateBody": {
            "type": "object",
            "properties": {
                "source": {
                    "$ref": "#/definitions/memory.SourceKind"
                },
                "text": {
                    "type": "string"
                }
            }
        },
        "memory.FactualInput": {
            "type": "object",
            "properties": {
                "account_id": {
                    "type": "string"
                },
                "agent_id": {
                    "type": "string"
                },
                "inputs": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/memory.InputItem"
                    }
                },
                "session_id": {
                    "type": "string"
                }
            }
        },
        "memory.InputItem": {
            "type": "object",
            "properties": {
                "content": {
                    "description": "Text as string, files as base64.",
                    "type": "string"
                },
                "content_type": {
                    "description": "MIME type: text/plain, application/pdf, image/png, etc.",
                    "type": "string"
                },
                "kind": {
                    "$ref": "#/definitions/memory.SourceKind"
                },
                "metadata": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                }
            }
        },
        "memory.MemoryInput": {
            "type": "object",
            "properties": {
                "account_id": {
                    "type": "string"
                },
                "agent_id": {
                    "type": "string"
                },
                "include_sources": {
                    "description": "When true, return original source content with facts.",
                    "type": "boolean"
                },
                "inputs": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/memory.InputItem"
                    }
                },
                "message_history": {
                    "description": "Number of recent raw messages to return. 0 = facts only.",
                    "type": "integer"
                },
                "session_id": {
                    "type": "string"
                }
            }
        },
        "memory.MemoryOutput": {
            "type": "object",
            "properties": {
                "facts": {
                    "description": "Relevant facts for the agent.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/memory.ReturnedFact"
                    }
                },
                "messages": {
                    "description": "Recent USER/AGENT messages from sources. Empty if MessageHistory is 0.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/memory.ConversationMessage"
                    }
                }
            }
        },
        "memory.ReturnedFact": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "kind": {
                    "$ref": "#/definitions/memory.FactKind"
                },
                "original_source": {
                    "description": "Populated only if IncludeSources is true.",
                    "type": "string"
                },
                "source_kind": {
                    "$ref": "#/definitions/memory.SourceKind"
                },
                "text": {
                    "type": "string"
                }
            }
        },
...
        "memory.SourceKind": {
            "type": "string",
            "enum": [
                "SYSTEM",
                "USER",
                "AGENT",
                "TOOL",
                "DOCUMENT",
                "CODE"
            ],
            "x-enum-comments": {
                "SOURCE_AGENT": "The agent's own generated output.",
                "SOURCE_CODE": "Source code, configuration files, schemas, technical artifacts.",
                "SOURCE_DOCUMENT": "Any non-code reference material. PDFs, text files, spreadsheets.",
                "SOURCE_SYSTEM": "Current date, session info, environment variables. Immutable.",
                "SOURCE_TOOL": "Output from any tool or sub-agent invoked by the agent.",
                "SOURCE_USER": "Input from whoever is initiating the request to the agent."
            },
            "x-enum-descriptions": [
                "Current date, session info, environment variables. Immutable.",
                "Input from whoever is initiating the request to the agent.",
                "The agent's own generated output.",
                "Output from any tool or sub-agent invoked by the agent.",
                "Any non-code reference material. PDFs, text files, spreadsheets.",
                "Source code, configuration files, schemas, technical artifacts."
            ],
            "x-enum-varnames": [
                "SOURCE_SYSTEM",
                "SOURCE_USER",
                "SOURCE_AGENT",
                "SOURCE_TOOL",
                "SOURCE_DOCUMENT",
                "SOURCE_CODE"
            ]
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}